'Refactored by Sourcery'

+++ api/chalicelib/core/sessions_favorite_viewed.py
@@ -5,13 +5,12 @@ from chalicelib.core import sessions
 def add_favorite_session(project_id, user_id, session_id):
     with pg_client.PostgresClient() as cur:
         cur.execute(
-            cur.mogrify(f"""\
-                INSERT INTO public.user_favorite_sessions 
-                    (user_id, session_id) 
-                VALUES 
-                    (%(userId)s,%(sessionId)s);""",
-                        {"userId": user_id, "sessionId": session_id})
+            cur.mogrify(
+                '\\\x1f                INSERT INTO public.user_favorite_sessions \x1f                    (user_id, session_id) \x1f                VALUES \x1f                    (%(userId)s,%(sessionId)s);',
+                {"userId": user_id, "sessionId": session_id},
+            )
         )
+
     return sessions.get_by_id2_pg(project_id=project_id, session_id=session_id, user_id=user_id, full_data=False,
                                   include_fav_viewed=True)
 
@@ -19,13 +18,12 @@ def add_favorite_session(project_id, user_id, session_id):
 def remove_favorite_session(project_id, user_id, session_id):
     with pg_client.PostgresClient() as cur:
         cur.execute(
-            cur.mogrify(f"""\
-                        DELETE FROM public.user_favorite_sessions                          
-                        WHERE 
-                            user_id = %(userId)s
-                            AND session_id = %(sessionId)s;""",
-                        {"userId": user_id, "sessionId": session_id})
+            cur.mogrify(
+                '\\\x1f                        DELETE FROM public.user_favorite_sessions                          \x1f                        WHERE \x1f                            user_id = %(userId)s\x1f                            AND session_id = %(sessionId)s;',
+                {"userId": user_id, "sessionId": session_id},
+            )
         )
+
     return sessions.get_by_id2_pg(project_id=project_id, session_id=session_id, user_id=user_id, full_data=False,
                                   include_fav_viewed=True)
